# Cache unitate function