@icambron 2023-11-12
@moh-incom This seems reasonable, but would be nice if the benchmark was included in the PR (see the benchmarks dir)