@moh-incom 2023-11-17
I was unaware of the `benchmarks` directory. Are there any instructions on running the benchmarks? The straight-forward guess `npm run benchmark` doesn't work for me (gives me a cryptic ESM/CommonJS incompatibility error)

From the looks of it, the benchmarks `DateTime.fromFormat` and `DateTime.fromFormat with zone` should already cover what my benchmark does.