@icambron 2023-11-18
I pushed up a fix for `npm run benchmark`