@moh-incom 2023-11-29
I found out a test from the test suite was failing because I was caching based on the `Locale.locale` field which does not contain the numbering system. I now cache on `Locale.intl` instead. This should cover all the problematic cases.

I got the benchmarks running after you fixed them. In addition, I added some more to more easily test the impact of my optimizations.

Even though it might be a bit outside the scope of this PR, I also added similar caching functionality to the `Locale.isEnglish` function. This speeds up `DateTime.fromFormat` incredibly when run with a non-English locale, which is also covered in one of the new benchmarks. If you prefer, I can move it to a different PR.